tlv: Remove Tag::Char and Tag::Bool
`char` is encoded in Unicode, and therefore is longer than originally
assumed (1 byte). This became apparent from tests, but could become an
issue if anyone used this expecting `tlv::Tag::Char` to actually be an
Unicode character.

Since arrays of every integer type is supported, each application may
convert its strings into a fitting representation, encode it into an array,
and encode that into a `tlv::Tag`.

Similarly, `tlv::Tag::Bool` was just an alias to sending `0 as u8` and
`1 as u8`, for false and true respectively.

Thus, remove `tlv::Tag::Char` and `tlv::Tag::Bool`.